nosql_porj/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sentence-transformers==3.2.1
qdrant-client==1.12.0
redis==5.1.1
clickhouse-connect==0.8.3
pymongo==4.10.1
python-dotenv==1.0.1
tqdm==4.66.5
pydantic==2.9.2
faker==30.3.0
